data-analysis/llm-data-quality/requirements_cpu.txt
pydantic==1.9.1
transformers==4.25.1
evaluate==0.4.0
tqdm==4.64.1
pyyaml==6.0
faiss-cpu